+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Support for Jellyfin's v12 (v12 stable release is set to deprecate 'X-Emby-Token' headers for API calls)
+
 ## [0.3.3] - 2026-07-24
 
 ### Fixed